Added changes for sync logic
