+++ tests/test_sync.py
@@ -13,9 +13,9 @@ except ImportError:
 
 class SyncIntegrationTest(TaboolaBaseTest, unittest.TestCase):
 
-    @patch("tap_taboola.singer.write_state")
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_state")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_sync_empty_campaigns_no_crash(
         self,
         mock_request,
@@ -35,9 +35,9 @@ class SyncIntegrationTest(TaboolaBaseTest, unittest.TestCase):
         ]
         self.assertEqual(len(campaign_records), 0)
 
-    @patch("tap_taboola.singer.write_state")
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_state")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_sync_empty_performance_no_crash(
         self,
         mock_request,
@@ -58,9 +58,9 @@ class SyncIntegrationTest(TaboolaBaseTest, unittest.TestCase):
         ]
         self.assertEqual(len(perf_records), 0)
 
-    @patch("tap_taboola.singer.write_state")
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_state")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_write_record_called_for_each_campaign(
         self,
         mock_request,
@@ -79,9 +79,9 @@ class SyncIntegrationTest(TaboolaBaseTest, unittest.TestCase):
         ]
         self.assertEqual(len(campaign_records), len(self.MOCK_CAMPAIGNS))
 
-    @patch("tap_taboola.singer.write_state")
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_state")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_write_record_called_for_each_performance_row(
         self,
         mock_request,
@@ -101,9 +101,9 @@ class SyncIntegrationTest(TaboolaBaseTest, unittest.TestCase):
         ]
         self.assertEqual(len(perf_records), len(self.MOCK_CAMPAIGN_PERFORMANCE))
 
-    @patch("tap_taboola.singer.write_state")
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_state")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_campaign_parse_types(
         self,
         mock_request,
@@ -122,9 +122,9 @@ class SyncIntegrationTest(TaboolaBaseTest, unittest.TestCase):
         self.assertIsInstance(record['is_active'], bool)
         self.assertIsInstance(record['name'], str)
 
-    @patch("tap_taboola.singer.write_state")
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_state")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_performance_parse_types(
         self,
         mock_request,
@@ -147,9 +147,9 @@ class SyncIntegrationTest(TaboolaBaseTest, unittest.TestCase):
         # Date should be YYYY-MM-DD format
         self.assertRegex(record['date'], r'^\d{4}-\d{2}-\d{2}$')
 
-    @patch("tap_taboola.singer.write_state")
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_state")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_campaign_null_end_date_uses_sentinel(
         self,
         mock_request,
