+++ tests/test_sync_canary.py
@@ -13,8 +13,8 @@ except ImportError:
 
 class SyncCanaryIntegrationTest(TaboolaBaseTest, unittest.TestCase):
 
-    @patch("tap_taboola.singer.write_record")
-    @patch("tap_taboola.request")
+    @patch("tap_taboola.streams.singer.write_record")
+    @patch("tap_taboola.streams.request")
     def test_full_pipeline_emits_records(
         self,
         mock_request,
